+++ pymhf/__init__.py
@@ -2,4 +2,4 @@ from .main import load_module  # noqa
 from .core.hooking import FuncHook  # noqa
 from .core.mod_loader import Mod, ModState  # noqa
 
-__version__ = "0.0.1"
+__version__ = "0.1.1"