+++ pymhf/core/__init__.py
@@ -4,4 +4,4 @@ from importlib.metadata import version, PackageNotFoundError
 try:
     __version__ = version("pymhf")
 except PackageNotFoundError:
-    __version__ = "0.6.12"
+    __version__ = "0.1.1"