+++ pymhf/core/_types.py
@@ -1,6 +1,6 @@
 from collections import namedtuple
 from enum import Enum
-from typing import Protocol, Any
+from typing import Any, Callable, Optional, Protocol
 
 
 FUNCDEF = namedtuple("FUNCDEF", ["restype", "argtypes"])
@@ -17,6 +17,7 @@ class HookProtocol(Protocol):
     _has__result_: bool
     _hook_func_name: str
     _hook_time: DetourTime
+    _custom_trigger: Optional[str]
 
     def __call__(self, *args: Any, **kwargs: Any) -> Any:
         ...