+++ pymhf/core/importing.py
@@ -1,16 +1,10 @@
-from abc import ABC
-from dataclasses import fields
-from functools import partial
-import inspect
 import importlib
 import importlib.util
-import json
 import logging
 import os.path as op
-import os
 import traceback
 from types import ModuleType
-from typing import Any, Optional, Callable
+from typing import Optional
 import string
 import sys